refactor: fix imports in components
